bug:3082865 don't use IN to construct sql to delete downloads
DownloadManager should not use IN clause when constructing
SQL to delete downloaded files Dowload app.
Lexer code in Download app doesn't know how to parse it.
Real fix is to fix Lexer.
But seriously real fix is to get rid of this complexity -
that I am planning for next version.

also, the following 2 are identical, in terms of SQL performance
   WHERE _id IN (?, ?, ?)
   WHERE (_id = ? OR _id = ? OR _id = ?)

Change-Id: Icca659a17c412247a193879e8d2f34e1b43ec9e5