Regenerate preloaded classes for Gingerbread.
Unlike previous releases where the set of classes-to-be-preloaded
was generated on a first-generation device (G1 and MyTouch 3G), this
round I used a second-generation device, a Nexus One. As a consequence,
class loading is faster (win) and fewer classes hit the 1250ms threshold.
Instead of sharing classes based on load time alone, classes are now shared
based on the number of applications loading them.

Change-Id: I18f7aa3e7e6258818871b3968b515c06314371b3
http://b/2546002