Show app services in adb bugreport (dumpstate).
There is a 60s timeout for this command.  I'd rather do it as an individual timeout
for each of the services, but this is easier and will probably cover for busted apps
well enough.  Also, this is running last in dumpstate, so if it's lost... oh well.

Also, clean up the formatting of the service dump so it's more readable in bugreport.

Change-Id: Ibd727f4b3366f7af9e7ef1e231868f5cedc48132