Hide plugin views when we start playing HTML5 video
Plugins like Flash draw directly to the screen on top
of all other views - including the fullscreen HTML5 video
view. This means that if you start an HTML5 video on a page
also serving a flash clip then flash will obscure the video.

This patch hides plugin views before start the video and shows
them again when the VideoView is destroyed.

Bug: 3183521
Change-Id: I9839c47e9d64b081651b952057e9283630362b29