DO NOT MERGE Dont let supplicant stay in DORMANT state
In some situations, the supplicant stays in a DORMANT
state. This prevents a connections to a network.

As long as the driver is started and we are not in scan-only
mode, issue a reconnect to get supplicant out of DORMANT
state.

Upon failures and network disable, supplicant should get
into an INACTIVE state.

Bug: 3236920
Change-Id: If32b5bfb7fea6c69c4a461db0b2012875b6273b5