change how assets are copied out of apk
Currently test code use getActivity to get DRT's assets. However
the side effect is that this will actually launch the activity.
This is not affecting any tests yet, however there are cleaner
ways to do this. The instrumentation framework provides a
getTargetContext call which will give access to app under test's
context, and therefore granting access to its assets.

Change-Id: I0560b0fa5681f80bcb9296beec3fd0549c40cc8e