scroll selected text in scrollable layers
Compute whether the initial tap point is in
a scrollable layer or not. If it is, get
the bounds of the scrollable layer and use them,
clipped to the view, to determine when the
select anchor is dragged close to the layer's
edge, to invoke autoscrolling.

Fix some bugs around stopping the autoscrolling
when there is nothing to do, and when a motion
up event is sent.

requires companion change in external/webkit

bug:3191699
Change-Id: Ic9229b063a8102ec93d0cd5bcaba774cc170c106