Fix 3303606: invalidate WaveView after resetting state
This fixes a bug where the unlock state machine was in the
reset state when it awakes which caused it to ignore
the first unlock attempt.  The reason for this is the state
machine didn't run again until the next user input.

We now invalidate the drawing area to ensure the unlock state
machine runs and is ready when LockScreen shows again.

Change-Id: I9f2b279d432ff634ba0b29e606cdc954718a2ad2