Fix another part of #3387855: Menu flashes on left side of screen
When task switching while dialog windows were up, the exit animation
wouldn't be seen.  The animation needs to force Z-order its windows
to the top.

Also fix a bug where the lock screen fade-out was immediately
Z-ordering the wallpaper to the bottom if the activity behind was
not on top of the wallpaper.

Change-Id: Icd45cd8ccfff9829822c1682931413246eb9601b